Removing old View based layout & updating test data
Removing the RecyclerViewAdapter and ViewHolder layout since this is handled with Compose now
Updating theme properties
Adding code to fill the width of the area it's in.
Renaming BillComposable to BillOverview because I don't think composable necessarily needs to be in the name, and this one is more representative of what it actually is.