Adding Add Bill Composable
Adding Add Bill Composable to allow users to add bills they want to track. Adding UI State and ViewModel to go with this. Exposed composable is stateful to populate a stateless version which is private.
To target API 31, exposing MainActivity in manifest with exported="true"
Doing some restructuring of packages in the project
Adding an Unspecified BillType to default to for the BillObject class.
Converting BillDetailsFragment to be fully compose based. Removing a lot of logic from the ViewModel which will just be included with BillObjects. Removing BillDetailRecyclerViewAdapter since the composables will render out exactly what we need on the screen. Moving a lot of Bill Detail composables into the BillDetails composable file. Since these elements are used exclusively inside the BillDetails composable, this consolidates the amount of files we have.
Making BillListFragment compose based.
Adding defaults to BillObject so this can be used inside of State Holders with an empty populated state.
Making amount inside of BillPayment private so the exposed property is a string and is already formatted how it is needed in the UI.
Moving all compose colors into an object for better organization.
Adding versions.gradle to have a dedicated area for keeping versions of libraries used in app.