Removing BillDetailFragment and Adding Compose Navigation
Removing Fragment from AddBillScreen file. Since all navigation is to other composables, we don't need to be instantiating Fragments. AppCompatActivity can be replaced with ComponentActivity so Composables can be used from the base of the activity.
Removing BillDetailFragment and replacing it with a composable that represents the same layout. Reusing BillDetailViewModel for use with the composable. Bill Details will start in a loading state. Once the bill has been fetched, we can update the state and load the data.
Removing references to BillTrackerTheme in composables since the base of the app is wrapped in the theme.
Moving CreditCardLimit from BillHistoryDetail composable. These are separate things and should render separate from each other.
Removing BillListFragment from app since a composable represents the same layout.
Adding ID to bills so they can be identified easily.
Adding Payment History to fake bill objects
Adding Hilt Navigation Compose to app so ViewModels can be instantiated with the NavGraph
Adding hilt_nav_compose version and bumping version of navigation to latest